requirements.txt
2captcha-python==1.2.0
absl-py==1.4.0
alabaster==0.7.13
allure-pytest==2.13.2
allure-python-commons==2.13.2
altgraph==0.17.3
Appium-Python-Client==2.9.0
apturl==0.5.2
argcomplete==1.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
Babel==2.12.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bitstring==4.0.2
blinker==1.4
bokeh==3.1.1
Brlapi==0.8.3
cachetools==5.3.1
certifi==2023.5.7
chainmap==1.0.3
chardet==4.0.0
click==8.0.3
colorama==0.4.4
combomethod==1.0.12
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
css-html-js-minify==2.5.5
cupshelpers==1.0
cycler==0.11.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docutils==0.18.1
duplicity==0.8.21
ecdsa==0.18.0
editdistance==0.6.2
efficientnet==1.0.0
esptool==4.5.1
essential-generators==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fasteners==0.14.1
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.39.4
furo==2023.5.20
future==0.18.2
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gyp==0.1
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.31.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jax==0.4.11
jeepney==0.7.1
Jinja2==3.1.2
kaleido==0.2.1
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
keyring==23.5.0
kiwisolver==1.4.4
knockpy==5.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
libclang==16.0.0
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==1.1.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdformat==0.7.16
mdformat_deflist==0.1.2
mdformat_frontmatter==2.0.1
mdformat_myst==0.1.5
mdformat_tables==0.4.1
mdit-py-plugins==0.3.5
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
myst-parser==1.0.0
netifaces==0.11.0
networkx==3.1
nulltype==2.3.1
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
options==1.4.10
outcome==1.2.0
packaging==23.1
pandas==2.0.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
plotly==5.14.1
pluggy==1.0.0
protobuf==4.23.2
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pyclipper==1.3.0.post4
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.42.1
pyinotify==0.9.6
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.3.1
pytest-check==2.1.4
pytest-rerunfailures==11.1.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
reedsolo==1.6.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
roman==3.3
rsa==4.9
rst-to-myst==0.3.4
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scipy==1.10.1
sdkmanager==0.5.1
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.9.1
shapely==2.0.1
six==1.12.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==6.2.1
sphinx-autobuild==2021.3.14
sphinx-basic-ng==1.0.0b1
sphinx-material==0.0.35
sphinx-rtd-dark-mode==1.2.4
sphinx-rtd-theme==1.2.0
sphinx-togglebutton==0.3.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
systemd-python==234
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
text-unidecode==1.3
tifffile==2023.4.12
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
TwoCaptcha==0.0.1
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.6
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
wadllib==1.3.6
webdriver-manager==3.8.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xyzservices==2023.5.0
zipp==1.0.0
